General Update: 			Include Windows running environment 			Length of File in Windows 1 Larger 			( 31 -> 32, 106 -> 107)
+ FDGD:			overlap should only be considered if > 0
+ InputData:			new Windows Path variables
+ Main:			New Directory Create System
+ Optimization:			TopNest allocation after NoTop variable declaration
+ ReadData:			Add createDirectoriesWindows
+ Toolbox:			Align precision values

+++ AerOpt_PhD/Toolbox.f90
@@ -101,9 +101,9 @@
     ! Local Variables
     integer :: left, right
     real :: random
-    real :: pivot
+    double precision :: pivot
     integer :: marker
-    real :: temp
+    double precision :: temp
 
     if (nA > 1) then
 
@@ -160,9 +160,9 @@
     ! Local Variables
     integer :: left, right
     real :: random
-    real :: pivot
+    double precision :: pivot
     integer :: marker
-    real :: temp
+    double precision :: temp
 
     if (nA > 1) then
